Blend search header into surface via shared HeaderGradient
The "Fact Check" / "Artikel-Suche" header painted a solid background-colored
block, which read as a hard white rectangle over the pale-blue surface and
looked inconsistent with the other headers (e.g. the personal tab).

Extract the personal header's fade recipe (app background color -> near
transparent) from AnimatedHeader into a shared HeaderGradient component and
use it in both places, so the search header blends into the surface below
like the rest of the app. Update the SearchHeader test to assert the new
gradient-based header instead of the old solid fill.
